Added creation of tar archive to TarImpl class
We now can create an archive based on our new tar implmentation class
though we still need to provide it all the files at once or we will need
to keep the file handles open which (currently) is not desireable. If we
try to do atomic adds on a closed file, we can only add uncompressed
content which we also don't want.