Change the delta diffing to be pipelined (not parallel yet)
This change moves the diffing process to use a pipelined approach which
is only a step away from being threaded now. Minor fix for directory
handling has been done to archive implementations to support this
changeover. Currently outstanding things that need to be done include
zip and tar handling of folders.