Now moved instantiation of archive object to a helper class
This change moves the abstractions from the patcher to a helper class so
we now use introspections to figure out what available archive
implementation classes we have. this change also allows for cleaner code
on 'with' blocks and removes any knowledge of undelying system from th
patcher while retaining the same functionality.