Fix problems with tar archive handling of folders
There were problems if a folder was in the archive and had permissions
and stat data that needed to be transfered over to the target dir since
the "expand" would always expand the whole directory. Another problem
would be encountered if the folder in the archive had a matching
filename as a regular file which is also fixed with this change.